Crée un composant générique HeaderFixed à partir de la logique présente dans ActionHeader
